social/requirements.txt
click>=8.1.0
tweepy>=4.14.0
atproto>=0.0.55
anthropic>=0.42.0
python-dotenv>=1.0.0
